fix: apply nicknames to thread mentions and rewrite room avatar urls
