+++ src/app/utils/matrix.ts
@@ -1,6 +1,6 @@
 import type { EncryptedAttachmentInfo } from 'browser-encrypt-attachment';
 import { decryptAttachment } from 'browser-encrypt-attachment';
-import { Channel, convertFileSrc, invoke, isTauri } from '@tauri-apps/api/core';
+import { Channel, invoke, isTauri } from '@tauri-apps/api/core';
 import type {
   AccountDataEvents,
   EventTimelineSet,
@@ -28,21 +28,11 @@ import { encryptAttachmentStreaming } from '$utils/attachmentCrypto';
 import { getEventReactions, getStateEvent } from './room';
 import { getReactionContent } from './messageReaction';
 import { matchMxId, validMxId } from './mxIdHelper';
-import {
-  fetchMediaBlob,
-  getCurrentMediaSessionScope,
-  type MediaTransportOptions,
-} from './mediaTransport';
+
+export { mxcUrlToHttp, rewriteAuthenticatedMediaUrl } from './mediaUrl';
+import { fetchMediaBlob, type MediaTransportOptions } from './mediaTransport';
 
 const DOMAIN_REGEX = /\b(?:[a-zA-Z0-9-]+\.)+[a-zA-Z]{2,}\b/;
-const TAURI_MEDIA_CACHE_VERSION = '__sable_media_cache=3';
-const TAURI_MEDIA_PATH_PREFIXES = [
-  '/_matrix/client/v1/media/',
-  '/_matrix/media/v3/download/',
-  '/_matrix/media/v3/thumbnail/',
-  '/_matrix/media/r0/download/',
-  '/_matrix/media/r0/thumbnail/',
-];
 
 export const isServerName = (serverName: string): boolean => DOMAIN_REGEX.test(serverName);
 
@@ -480,60 +470,6 @@ export const removeRoomIdFromMDirect = async (mx: MatrixClient, roomId: string):
     EventType.Direct as string as unknown as keyof AccountDataEvents,
     userIdToRoomIds
   );
-};
-
-export const rewriteAuthenticatedMediaUrl = (httpUrl: string | null): string | null => {
-  if (!httpUrl) return null;
-  if (!isTauri()) return httpUrl;
-  const sourceUrl = httpUrl.startsWith('sable-media://')
-    ? httpUrl.slice('sable-media://'.length)
-    : httpUrl;
-  let parsedUrl: URL;
-  try {
-    parsedUrl = new URL(sourceUrl);
-  } catch {
-    return httpUrl;
-  }
-  if (
-    (parsedUrl.protocol !== 'http:' && parsedUrl.protocol !== 'https:') ||
-    parsedUrl.origin === 'null' ||
-    !TAURI_MEDIA_PATH_PREFIXES.some((path) => parsedUrl.pathname.startsWith(path))
-  ) {
-    return httpUrl;
-  }
-  if (httpUrl.includes(TAURI_MEDIA_CACHE_VERSION)) return httpUrl;
-  const mediaUrl = httpUrl.startsWith('sable-media://')
-    ? httpUrl
-    : convertFileSrc(httpUrl, 'sable-media');
-  // Session-scoped so the cacheable response is never shared across accounts.
-  const sessionScope = encodeURIComponent(getCurrentMediaSessionScope());
-  const separator = mediaUrl.includes('?') ? '&' : '?';
-  return `${mediaUrl}${separator}${TAURI_MEDIA_CACHE_VERSION}&__sable_media_session=${sessionScope}`;
-};
-
-export const mxcUrlToHttp = (
-  mx: MatrixClient,
-  mxcUrl: string,
-  useAuthentication?: boolean,
-  width?: number,
-  height?: number,
-  resizeMethod?: string,
-  allowDirectLinks?: boolean
-): string | null => {
-  const httpUrl = mx.mxcUrlToHttp(
-    mxcUrl.replace(/^["']|["']$/g, ''),
-    width,
-    height,
-    resizeMethod,
-    allowDirectLinks,
-    undefined,
-    useAuthentication
-  );
-
-  if (httpUrl && isTauri()) {
-    return rewriteAuthenticatedMediaUrl(httpUrl);
-  }
-  return httpUrl;
 };
 
 export const downloadMedia = async (src: string, options?: MediaTransportOptions): Promise<Blob> =>
